refactor(tests): split oversized pre-PR validation test module (#4352)
tests/test_validation_pre_pr.py was 1511 lines with 106 tests, well
above the 500-line taste-lint ceiling. Read PR #4341 and issue #4285
before splitting: #4341 rewrites pre_pr_sequence.py (a different file)
and #4285 concluded that extraction was wrong for the data module.
The test file split is independent and correct.

Split into four files, all under 500 lines:
- tests/test_validation_pre_pr.py (443 lines): core runner, parser,
  session log, design review, main entry point
- tests/test_validation_pre_pr_markdown.py (477 lines): markdown lint
  and dash prohibition tests
- tests/test_validation_pre_pr_environment.py (344 lines): lefthook,
  worktree, workflow-YAML tests
- tests/test_validation_pre_pr_gate_validators.py (262 lines): vendor
  portability, review marker, CI pin tests

Before: 106 tests collected. After: 106 tests collected.
Taste-lint baseline updated from 598 to 597.

Fixes #4352

Copilot-Session: fd14799b-742e-4bc2-88bf-3d673ab259f4
